+++ libevmasm/AssemblyItem.cpp
@@ -91,6 +91,8 @@ size_t AssemblyItem::bytesRequired(size_t _addressLength) const
 			return 1 + (3 + 32) * *m_immutableOccurrences;
 		else
 			return 1 + (3 + 32) * 1024; // 1024 occurrences are beyond the maximum code size anyways.
+	case VerbatimBytecode:
+		return std::get<2>(*m_verbatimBytecode).size();
 	default:
 		break;
 	}
@@ -101,6 +103,8 @@ size_t AssemblyItem::arguments() const
 {
 	if (type() == Operation)
 		return static_cast<size_t>(instructionInfo(instruction()).args);
+	else if (type() == VerbatimBytecode)
+		return get<0>(*m_verbatimBytecode);
 	else if (type() == AssignImmutable)
 		return 2;
 	else
@@ -126,6 +130,8 @@ size_t AssemblyItem::returnValues() const
 		return 1;
 	case Tag:
 		return 0;
+	case VerbatimBytecode:
+		return get<1>(*m_verbatimBytecode);
 	default:
 		break;
 	}
@@ -241,6 +247,9 @@ string AssemblyItem::toAssemblyText(Assembly const& _assembly) const
 	case UndefinedItem:
 		assertThrow(false, AssemblyException, "Invalid assembly item.");
 		break;
+	case VerbatimBytecode:
+		text = string("verbatimbytecode_") + util::toHex(get<2>(*m_verbatimBytecode));
+		break;
 	default:
 		assertThrow(false, InvalidOpcode, "");
 	}
@@ -308,6 +317,9 @@ ostream& solidity::evmasm::operator<<(ostream& _out, AssemblyItem const& _item)
 		break;
 	case AssignImmutable:
 		_out << " AssignImmutable";
+		break;
+	case VerbatimBytecode:
+		_out << " Verbatim " << util::toHex(_item.verbatimData());
 		break;
 	case UndefinedItem:
 		_out << " ???";